fixed a parsing issue
The OppositeBandInformation::Answer and the INFORMATION::Answer functions incorrectly parsed the Mode value. When the radio was set to modes above FM the hexadecimal value would throw an stoi exception. I gave OppositeBandInformation it's own type OppositeInformationValue so one can display the mode value and frequency for VFOB when A/B is pressed on the rig.  InformationValue is for VFOA, OppositeInformationValue is for VFOB.