updated to add a dialog for connecting to the radio
Rerwote the UI to use lazy loading to not exceed the rs232 time out limits of the radio. Added a simple dialog that will enable users to enter the serial port (be it a COM port on windows or a /dev/tty on linux) and select the baud rate of the radio. Added controls for the CAT settings of the radio. Build another testing version for the testers to try out and provide feedback.