Update to the Control program
Added a memory channel editor, fixed a few bugs in the controller and library that I discovered while writing the memory editor for the control program. Now there's all but a few items left on the aspects of the operation of the FT-891 that's not directly accesable in this application.